update: merge background services into one class Because when registering more than 2 services, some anomalies happen with the stopping token, causing them to crash the app, combining all workers into 1 method solves the problem
