+++ src/Api/ConfigureServices.cs
@@ -18,7 +18,7 @@ public static class ConfigureServices
         
         // Register services
         services.AddServices();
-        services.AddHostedService<ExpiryPermissionService>();
+        services.AddHostedService<BackgroundWorkers>();
         
         services.AddControllers(opt =>
             opt.Conventions.Add(new RouteTokenTransformerConvention(new SlugifyParameterTransformer())));
